+++ multiqc/modules/cutadapt/cutadapt.py
@@ -133,7 +133,7 @@ class MultiqcModule(BaseMultiqcModule):
                 if s_name == "-":
                     s_name = f["s_name"]
                 else:
-                    s_name = self.clean_s_name(s_name, f["root"])
+                    s_name = self.clean_s_name(s_name, f)
                 if s_name in self.cutadapt_data:
                     log.debug("Duplicate sample name found! Overwriting: {}".format(s_name))
                 self.cutadapt_data[s_name] = dict()